+++ genomix/genomix-driver/src/main/java/edu/uci/ics/genomix/mixture/model/FittingMixture.java
@@ -75,60 +75,60 @@ public class FittingMixture {
         outstream.close();
     }
 
-    public double computeLikelihood(float[] data, float prob_Exp, float prob_Normal, float[] prob_expMembership,
-            float[] prob_normalMembership) {
-        float likelihood = 0;
+    public double computeLikelihood(double[] data, double prob_Exp, double prob_Normal, double[] prob_expMembership,
+            double[] prob_normalMembership) {
+        double likelihood = 0;
         for (int i = 0; i < data.length; i++)
             likelihood += Math.log(prob_Exp * prob_expMembership[i] + prob_Normal * prob_normalMembership[i]);
         System.out.println(likelihood);
         return likelihood;
     }
 
-    public static double fittingMixture(float[] data, float max, int numOfIterations) throws IOException {
+    public static double fittingMixture(double[] data, double max, int numOfIterations) throws IOException {
         FittingMixture fm = new FittingMixture();
         ArrayList<Double> likelihoods = new ArrayList<Double>();
 
         /* initial guess at parameters */
-        float cur_expMean = 5;
+        double cur_expMean = 5;
         ExponentialDistribution cur_expDist = new ExponentialDistribution(cur_expMean);
 
-        float cur_normalMean = 20;
-        float cur_normalStd = 5;
+        double cur_normalMean = 20;
+        double cur_normalStd = 5;
         NormalDistribution cur_normalDist = new NormalDistribution(cur_normalMean, cur_normalStd);
 
         /* probability of each data point belonging to each class */
-        float prob_Exp = 0.5f;
-        float prob_Normal = 0.5f;
+        double prob_Exp = 0.5;
+        double prob_Normal = 0.5;
 
         // initiate prob_membership of exp and normal
-        float[] prob_expMembership = new float[data.length];
-        float[] prob_normalMembership = new float[data.length];
+        double[] prob_expMembership = new double[data.length];
+        double[] prob_normalMembership = new double[data.length];
 
         //        likelihoods.add(computeLikelihood(data, prob_Exp, prob_Normal, prob_expMembership, prob_normalMembership));
         /* Given a fixed set of parameters, it chooses the probability distribution 
          * which maximizes likelihood */
         for (int i = 0; i < data.length; i++) {
             // calculate the expectation prob of the two classes
-            prob_expMembership[i] = (float) cur_expDist.density(data[i]) * prob_Exp;
+            prob_expMembership[i] = cur_expDist.density(data[i]) * prob_Exp;
             if (prob_expMembership[i] == 0)
-                prob_expMembership[i] = 0.000000001f;
-            prob_normalMembership[i] = (float) cur_normalDist.density(data[i]) * prob_Normal;
+                prob_expMembership[i] = 0.000000001;
+            prob_normalMembership[i] = cur_normalDist.density(data[i]) * prob_Normal;
             if (prob_normalMembership[i] == 0)
-                prob_normalMembership[i] = 0.000000001f;
+                prob_normalMembership[i] = 0.000000001;
 
             // normalize
-            float membershipSum = prob_expMembership[i] + prob_normalMembership[i];
+            double membershipSum = prob_expMembership[i] + prob_normalMembership[i];
             prob_expMembership[i] = prob_expMembership[i] / membershipSum;
             prob_normalMembership[i] = prob_normalMembership[i] / membershipSum;
         }
-        float exp_overallSum = 0;
-        float normal_overallSum = 0;
+        double exp_overallSum = 0;
+        double normal_overallSum = 0;
         for (int i = 0; i < data.length; i++) {
             exp_overallSum += prob_expMembership[i];
             normal_overallSum += prob_normalMembership[i];
         }
         // normalize
-        float overallSum = exp_overallSum + normal_overallSum;
+        double overallSum = exp_overallSum + normal_overallSum;
         prob_Exp = exp_overallSum / overallSum;
         prob_Normal = normal_overallSum / overallSum;
 
@@ -150,7 +150,7 @@ public class FittingMixture {
                 cur_normalStd += prob_normalMembership[i] * Math.pow(data[i] - cur_normalMean, 2);
             }
 
-            cur_normalStd = (float) Math.sqrt(cur_normalStd / normal_overallSum);
+            cur_normalStd = Math.sqrt(cur_normalStd / normal_overallSum);
             if (cur_normalStd == 0)
                 cur_normalStd = 0.000000001f;
             cur_normalDist = new NormalDistribution(cur_normalMean, cur_normalStd);
@@ -159,15 +159,15 @@ public class FittingMixture {
              * which maximizes likelihood */
             for (int i = 0; i < data.length; i++) {
                 // calculate the expectation prob of the two classes
-                prob_expMembership[i] = (float) cur_expDist.density(data[i]) * prob_Exp;
+                prob_expMembership[i] = cur_expDist.density(data[i]) * prob_Exp;
                 if (prob_expMembership[i] == 0)
-                    prob_expMembership[i] = 0.000000001f;
-                prob_normalMembership[i] = (float) cur_normalDist.density(data[i]) * prob_Normal;
+                    prob_expMembership[i] = 0.000000001;
+                prob_normalMembership[i] = cur_normalDist.density(data[i]) * prob_Normal;
                 if (prob_normalMembership[i] == 0)
-                    prob_normalMembership[i] = 0.000000001f;
+                    prob_normalMembership[i] = 0.000000001;
 
                 // normalize
-                float membershipSum = prob_expMembership[i] + prob_normalMembership[i];
+                double membershipSum = prob_expMembership[i] + prob_normalMembership[i];
                 prob_expMembership[i] = prob_expMembership[i] / membershipSum;
                 prob_normalMembership[i] = prob_normalMembership[i] / membershipSum;
             }
@@ -185,7 +185,7 @@ public class FittingMixture {
             likelihoods.add(fm
                     .computeLikelihood(data, prob_Exp, prob_Normal, prob_expMembership, prob_normalMembership));
         }
-        
+
         /* For test */
         //        double[] exp_randomSample = cur_expDist.sample(5000);
         //        Map<Double, Long> exp_histData = fm.calcHistogram(exp_randomSample, 0, max, (int) (max));
@@ -201,15 +201,15 @@ public class FittingMixture {
         //
         //        // plot likelihood's change
         //        fm.drawHist(likelihoods, "Likelihood");
-        
+
         /* one choice of threshold is when the probability of belonging to the normal distribution 
          is greater than the probability of belonging to the exponential distribution
          this happens when the ratio of norm to exponential is > 1 */
-        float prob_in_exp = 0;
-        float prob_in_normal = 0;
-        for (float cov = 1; cov < max; cov++) {
-            prob_in_exp = prob_Exp * (float) cur_expDist.density(cov);
-            prob_in_normal = prob_Normal * (float) cur_normalDist.density(cov);
+        double prob_in_exp = 0;
+        double prob_in_normal = 0;
+        for (double cov = 1; cov < max; cov++) {
+            prob_in_exp = prob_Exp * cur_expDist.density(cov);
+            prob_in_normal = prob_Normal * cur_normalDist.density(cov);
             if (prob_in_exp < prob_in_normal)
                 return cov;
         }
